+++ packages/atomic/src/components/commerce/atomic-commerce-query-summary/e2e/atomic-commerce-query-summary.e2e.ts
@@ -25,12 +25,15 @@ test.describe('after searching for kayak', () => {
 });
 
 test.describe('when search yields no products', () => {
-  test.beforeEach(async ({querySummary}) => {
-    await querySummary.load({story: 'no-products'});
+  test.beforeEach(async ({querySummary, page}) => {
+    querySummary.noProducts();
+    await querySummary.load({story: 'default '});
+    await page.waitForLoadState('networkidle');
   });
 
   test('should not display anything', async ({querySummary}) => {
-    await expect(querySummary.hydrated).toBeEmpty();
+    await expect(querySummary.placeholder).not.toBeVisible();
+    await expect(querySummary.container).not.toBeVisible();
   });
 });
 
@@ -46,20 +49,31 @@ test.describe('when search yields multiple products', () => {
 });
 
 test.describe('when a query yield a single product', () => {
-  test.beforeEach(async ({querySummary, searchBox}) => {
+  test.beforeEach(async ({querySummary, searchBox, page}) => {
+    page.route('**/commerce/v2/search', async (route) => {
+      const response = await route.fetch();
+      const body = await response.json();
+      body.products = [body.products[0]];
+      body.pagination.totalEntries = 1;
+      body.pagination.totalPages = 1;
+      await route.fulfill({
+        response,
+        json: body,
+      });
+    });
     await querySummary.load({story: 'with-search-box'});
     await searchBox.hydrated.waitFor();
-    await searchBox.searchInput.fill('@ec_product_id=SP03730_00007');
+    await searchBox.searchInput.fill('kayak');
     await searchBox.submitButton.click();
   });
 
   test('should display message', async ({querySummary}) => {
-    const textRegex = /^Product 1 of 1 for @ec_product_id=SP03730_00007$/;
+    const textRegex = /^Product 1 of 1 for kayak$/;
     await expect(querySummary.text(textRegex)).toBeVisible();
   });
 
   test('screen readers should read out', async ({querySummary}) => {
-    const textRegex = /^Product 1 of 1 for @ec_product_id=SP03730_00007$/;
+    const textRegex = /^Product 1 of 1 for kayak$/;
     await expect(querySummary.ariaLive(textRegex)).toBeVisible();
   });
 });