+++ packages/atomic/src/components/search/facets/atomic-category-facet/atomic-category-facet.new.stories.tsx
@@ -9,11 +9,19 @@ const meta: Meta = {
   component: 'atomic-category-facet',
   title: 'Atomic/CategoryFacet',
   id: 'atomic-category-facet',
-
   render: renderComponent,
   decorators: [decorator],
   parameters,
   play,
+  argTypes: {
+    'attributes-number-of-values': {
+      name: 'number-of-values',
+      control: {type: 'number', min: 1},
+    },
+  },
+  args: {
+    'attributes-number-of-values': 8,
+  },
 };
 
 export default meta;